update test files for states
